+++ excalidraw-journals.js
@@ -23,6 +23,7 @@ function _excaMakeRoomCypher() {
 function _excaDelayedClose(app, fallback) {
   app.element.find('.window-header')[0].innerHTML = 'Saving...';
   app.minimize();
+  app.maximize = () => {};
   setTimeout(() => {fallback()}, 20000);
 }
 
@@ -70,31 +71,120 @@ class ExcalidrawActor extends ActorSheet {
 
 }
 
+class ExcalidrawJournal extends JournalSheet {
+
+  _getHeaderButtons() {
+    const buttons = super._getHeaderButtons()
+      .filter(b => b.label === 'Close');
+    if (game.user.isGM) {
+      buttons.unshift({
+        label: "Journal",
+        class: "entry-text",
+        icon: "fas fa-book",
+        onclick: async ev => {
+          const sheet = this.object.sheet;
+          await super.close()
+          this.object._sheet = null;
+          delete this.object.apps[sheet.appId];
+          await this.document.unsetFlag('core', 'sheetClass');
+          this.object.sheet.render(true);
+        }
+      })
+    }
+    return buttons;
+  }
+
+  close() {
+    _excaDelayedClose(this, () => super.close());
+  }
+
+  async getData() {
+    const _default = super.getData();
+    _excaAssignRoom(this.document);
+    return {
+      ..._default,
+      roomCode: this.document.roomCode,
+      roomCypher: this.document.roomCypher
+    };
+  }
+
+  static get defaultOptions() {
+    const _default = super.defaultOptions;
+
+    return {
+      ..._default,
+      classes: ['sheet', 'excalidraw-journal', 'form'],
+      width: canvas.screenDimensions[0] * 0.50,
+      height: canvas.screenDimensions[1] * 0.75,
+      resizable: true,
+      template: `./modules/excalidraw-journals/excalidraw-sheet.hbs`
+    };
+  }
+
+  get template() {
+    return `./modules/excalidraw-journals/excalidraw-sheet.hbs`;
+  }
+
+}
+
+class ExcalidrawItem extends ItemSheet {
+
+  close() {
+    _excaDelayedClose(this, () => super.close());
+  }
+
+  async getData() {
+    const _default = super.getData();
+    _excaAssignRoom(this.document);
+    return {
+      ..._default,
+      roomCode: this.document.roomCode,
+      roomCypher: this.document.roomCypher
+    };
+  }
+
+  static get defaultOptions() {
+    const _default = super.defaultOptions;
+
+    return {
+      ..._default,
+      classes: ['sheet', 'excalidraw-journal', 'form'],
+      width: canvas.screenDimensions[0] * 0.50,
+      height: canvas.screenDimensions[1] * 0.75,
+      resizable: true,
+      template: `./modules/excalidraw-journals/excalidraw-sheet.hbs`
+    };
+  }
+
+  get template() {
+    return `./modules/excalidraw-journals/excalidraw-sheet.hbs`;
+  }
+
+}
+
 // Adds Excalidraw as a Sheet Mode for Journal Entries
 Hooks.once('ready', function () {
   Actors.registerSheet('excalidraw-journals', ExcalidrawActor);
+  Journal.registerSheet('excalidraw-journals', ExcalidrawJournal);
+  Items.registerSheet('excalidraw-journals', ExcalidrawItem);
 
   Hooks.on('getJournalSheetHeaderButtons', function(app, buttons) {
-    if (app.constructor.name === 'JournalSheet') {
+    if (app.object._getSheetClass().name !== 'ExcalidrawJournal') {
       buttons.unshift({
         label: "Excalidraw",
         class: "entry-excalidraw",
         icon: "fas fa-pen-alt",
-        onclick: async ev => app._onSwapMode(ev, "excalidraw")
-      })
+        onclick: async ev => {
+          const sheet = app.object.sheet;
+          await sheet.close()
+          app.object._sheet = null;
+          delete app.object.apps[sheet.appId];
+          await app.document.setFlag('core', 'sheetClass', `excalidraw-journals.ExcalidrawJournal`);
+          app.object.sheet.render(true);
+        }
+      });
     }
   });
-
-  libWrapper.register('excalidraw-journals', 'JournalSheet.prototype.template', function (wrapped, ...args) {
-    if ( this._close) this.close = this._close;
-    if ( this._sheetMode === "image" ) return ImagePopout.defaultOptions.template
-    else if (this._sheetMode === "excalidraw") {
-      _excaAssignRoom(this.document);
-      this._close = this.close;
-      this.close = () => _excaDelayedClose(this, () => this._close());
-      return `./modules/excalidraw-journals/excalidraw-sheet.hbs`
-    } else return "templates/journal/sheet.html";
-  }, 'OVERRIDE');
 });
 
 // Adds Journal Note offline Excalidraw to left controls